Add files via upload
The evolution of distributed systems necessitated a standardized communication protocol. This led to the development of CORBA, or Common Object Request Broker Architecture, a middleware technology that has a significant role in the field of distributed computing. It enables communication between objects in a distributed environment regardless of their implementation details such as specific operating systems, programming languages, and hardware platforms. In this report, I will provide a comprehensive
discussion on aspects related to CORBA, following the outlines below: